Refactor: code review remediation — typed enums, DRY helpers, decomposed methods
9-phase remediation addressing 34 code review issues:

- Derive Copy on BuildingType/DisasterType, Eq on ActiveSiege/EventParticipant,
  RiverData::length usize→u32
- Replace stringly-typed fields with enums: GovernmentType, FeatureType,
  ResourceType, DerivationMethod; distortions String→Vec<Value>
- Consolidate duplicated helpers into sim/helpers.rs (is_food_resource,
  apply_stability_delta, faction_capital_oldest/largest, EPITHETS, etc.)
- Disease system: add extra key constants, use set_extra() for audit trail
- Scenario builder: #[must_use] on refs, SettlementRef auto-syncs population,
  default_for_kind takes EntityKind by value, remove add_from()
- Idiomatic patterns: expect→unwrap_or_else, manual shuffle→SliceRandom,
  unreachable! for Season, filter_map for sentinels, assert! for timestamps
- Extract ~50 magic numbers to named constants in knowledge.rs and buildings.rs
- Test quality: remove 13 get_* wrappers, fix tautological assert, use
  assert_deterministic()
- Decompose long methods: demographics tick(), politics/knowledge handle_signals()

+++ src/worldgen/rivers.rs
@@ -136,7 +136,7 @@ pub fn generate_rivers(world: &mut World, config: &WorldGenConfig, rng: &mut dyn
         let name = format!("{prefix} {suffix}");
 
         let region_path: Vec<u64> = path.iter().map(|&i| region_ids[i]).collect();
-        let river_length = path.len();
+        let river_length = path.len() as u32;
         let river_id = world.add_entity(
             EntityKind::River,
             name,
